fix resource alias
