+++ .github/workflows/ci.yml
@@ -16,7 +16,7 @@ on:
       - 'README.md'
     branches:
       - main
-      - fix-sweepeers
+      - fix-resource-alias
 
 # Ensures only 1 action runs per PR and previous is canceled on new trigger
 concurrency:
